+++ example/lib/src/no_route.dart
@@ -5,11 +5,11 @@ class NoRoute extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
         appBar: AppBar(
-          title: Text("can't find route"),
+          title: const Text('can\'t find route'),
         ),
         body: Center(
           child: Container(
-            child: Text("can't find route"),
+            child: const Text('can\'t find route'),
           ),
         ));
   }