+++ example/module_common/lib/module_common.dart
@@ -1,7 +1,0 @@
-library module_common;
-
-export 'src/at_text.dart';
-export 'src/dollar_text.dart';
-export 'src/emoji_text.dart';
-export 'src/my_extended_text_selection_controls.dart';
-export 'src/my_special_text_span_builder.dart';