+++ example/module_service/lib/module_service.dart
@@ -1,4 +1,0 @@
-library module_service;
-
-export 'src/tu_chong_repository.dart';
-export 'src/tu_chong_source.dart';