+++ example/module_widget/lib/module_widget.dart
@@ -1,6 +1,0 @@
-library module_widget;
-
-export 'src/crop_image.dart';
-export 'src/item_builder.dart';
-export 'src/pic_swiper.dart';
-export 'src/push_to_refresh_header.dart';